Refactor decorators
